Create checkout cart component
