+++ client/store/cart.js
@@ -86,7 +86,6 @@ export default function(state = defaultCart, action) {
       return action.cart
     case UPDATE_CART:
       return action.cart
-
     default:
       return state
   }
